+++ index.js
@@ -1,4 +1,11 @@
+const BinarySearchTreeNode = require('./src/binarySearchTreeNode');
+const AvlTreeNode = require('./src/binarySearchTreeNode');
 const BinarySearchTree = require('./src/binarySearchTree');
 const AvlTree = require('./src/avlTree');
 
-module.exports = { BinarySearchTree, AvlTree };
+module.exports = {
+  BinarySearchTreeNode,
+  BinarySearchTree,
+  AvlTreeNode,
+  AvlTree
+};