+++ src/avlTree.js
@@ -1,5 +1,5 @@
 /**
- * @datastructures-js/binary-search-tree
+ * datastructures-js/binary-search-tree
  * @copyright 2020 Eyas Ranjous <eyas.ranjous@gmail.com>
  * @license MIT
  */
@@ -13,154 +13,157 @@ const AvlTreeNode = require('./avlTreeNode');
  */
 class AvlTree extends BinarySearchTree {
   /**
+   * Applies the proper rotation on a node
    * @private
-   * applies the proper rotation on a node during insert or remove
    * @param {AvlTreeNode} node
    */
   _balanceNode(node) {
     if (!node) return;
 
     node.updateHeight();
-    const balance = node.calculateBalance();
+    const balance = node.getBalance();
     if (balance > 1) {
-      if (node.getLeft().getLeft() !== null) {
+      if (node.getLeft().hasLeft()) {
         node.rotateRight();
-      } else if (node.getLeft().getRight() !== null) {
+      } else if (node.getLeft().hasRight()) {
         node.rotateLeftRight();
       }
     } else if (balance < -1) {
-      if (node.getRight().getRight() !== null) {
+      if (node.getRight().hasRight()) {
         node.rotateLeft();
-      } else if (node.getRight().getLeft() !== null) {
+      } else if (node.getRight().hasLeft()) {
         node.rotateRightLeft();
       }
     }
-    if (node === this._root && (balance < -1 || balance > 1)) {
+
+    // check if root was rotated
+    if ((balance < -1 || balance > 1) && node === this._root) {
+      // replace root when rotated with the child (now parent of root)
       this._root = node.getParent();
     }
   }
 
   /**
-   * @public
-   * inserts a node with a key/value into tree
+   * Inserts a node with a key/value into tree
    * and maintains the tree balanced by applying the necessary rotations
+   *
+   * @public
    * @param {number|string} key
-   * @param {object} value
-   * @param {AvlTreeNode} node
-   * @return {AvlTreeNode} the inserted node
+   * @param {any} value
+   * @return {AvlTree}
    */
-  insert(key, value, node = this._root) {
-    if (node === null) {
-      this._root = new AvlTreeNode(key, value);
+  insert(key, value) {
+    const newNode = new AvlTreeNode(key, value);
+    const insertRecursive = (current) => {
+      if (key < current.getKey()) {
+        if (current.hasLeft()) {
+          insertRecursive(current.getLeft());
+          this._balanceNode(current); // backward-tracking
+        } else {
+          newNode.setParent(current);
+          current.setLeft(newNode).updateHeight();
+          this._count += 1;
+        }
+      } else if (key > current.getKey()) {
+        if (current.hasRight()) {
+          insertRecursive(current.getRight());
+          this._balanceNode(current); // backward-tracking
+        } else {
+          newNode.setParent(current);
+          current.setRight(newNode).updateHeight();
+          this._count += 1;
+        }
+      } else {
+        current.setValue(value);
+      }
+    };
+
+    if (this._root === null) {
+      this._root = newNode;
       this._count += 1;
-      return this._root;
+    } else {
+      insertRecursive(this._root);
     }
 
-    if (key < node.getKey() && node.getLeft() === null) {
-      const newNode = new AvlTreeNode(key, value);
-      node.setLeft(newNode);
-      newNode.setParent(node);
-      node.updateHeight();
-      this._count += 1;
-      return newNode;
-    }
-
-    if (key > node.getKey() && node.getRight() === null) {
-      const newNode = new AvlTreeNode(key, value);
-      node.setRight(newNode);
-      newNode.setParent(node);
-      node.updateHeight();
-      this._count += 1;
-      return newNode;
-    }
-
-    if (key === node.getKey()) {
-      node.setValue(value);
-      return node;
-    }
-
-    if (key < node.getKey()) {
-      const newNode = this.insert(key, value, node.getLeft());
-      this._balanceNode(node); // backward-tracking
-      return newNode;
-    }
-
-    const newNode = this.insert(key, value, node.getRight());
-    this._balanceNode(node); // backward-tracking
     return newNode;
   }
 
   /**
-   * @public
-   * remove a node by its key
+   * Removes a node by its key
    * and maintains the tree balanced by applying the necessary rotations
+   *
+   * @public
    * @param {number|string} key
-   * @param {AvlTreeNode} node
    * @return {boolean}
    */
-  remove(key, node = this._root) {
-    if (node === null) return false;
-
-    if (key < node.getKey()) {
-      const removed = this.remove(key, node.getLeft());
-      this._balanceNode(node);
-      return removed;
-    }
-
-    if (key > node.getKey()) {
-      const removed = this.remove(key, node.getRight());
-      this._balanceNode(node);
-      return removed;
-    }
-
-    if (node.getLeft() === null && node.getRight() === null) {
-      if (node.getParent() === null) {
-        this._root = null;
-      } else if (key < node.getParent().getKey()) {
-        node.getParent().setLeft(null);
-        node.getParent().updateHeight();
-      } else {
-        node.getParent().setRight(null);
-        node.getParent().updateHeight();
+  remove(key) {
+    const removeRecursively = (k, current) => {
+      if (current === null) {
+        return false;
       }
-      this._count -= 1;
-      return true;
-    }
 
-    if (node.getRight() === null) {
-      if (node.getParent() === null) {
-        this._root = node.getLeft();
-      } else if (key < node.getParent().getKey()) {
-        node.getParent().setLeft(node.getLeft());
-        node.getParent().updateHeight();
-      } else {
-        node.getParent().setRight(node.getLeft());
-        node.getParent().updateHeight();
+      if (k < current.getKey()) {
+        const removed = removeRecursively(k, current.getLeft());
+        this._balanceNode(current);
+        return removed;
       }
-      node.getLeft().setParent(node.getParent());
-      this._count -= 1;
-      return true;
-    }
 
-    if (node.getLeft() === null) {
-      if (node.getParent() === null) {
-        this._root = node.getRight();
-      } else if (key < node.getParent().getKey()) {
-        node.getParent().setLeft(node.getRight());
-        node.getParent().updateHeight();
-      } else {
-        node.getParent().setRight(node.getRight());
-        node.getParent().updateHeight();
+      if (k > current.getKey()) {
+        const removed = removeRecursively(k, current.getRight());
+        this._balanceNode(current);
+        return removed;
       }
-      node.getRight().setParent(node.getParent());
-      this._count -= 1;
-      return true;
-    }
 
-    const minRight = this.min(node.getRight());
-    node.setKey(minRight.getKey());
-    node.setValue(minRight.getValue());
-    return this.remove(minRight.getKey(), minRight);
+      // current node is the node to remove
+
+      // case 1: node has no children
+      if (current.isLeaf()) {
+        if (current.isRoot()) {
+          this._root = null;
+        } else if (k < current.getParent().getKey()) {
+          current.getParent().setLeft(null).updateHeight();
+        } else {
+          current.getParent().setRight(null).updateHeight();
+        }
+        this._count -= 1;
+        return true;
+      }
+
+      // case 2: node has a left child and no right child
+      if (!current.hasRight()) {
+        if (current.isRoot()) {
+          this._root = current.getLeft();
+        } else if (k < current.getParent().getKey()) {
+          current.getParent().setLeft(current.getLeft()).updateHeight();
+        } else {
+          current.getParent().setRight(current.getLeft()).updateHeight();
+        }
+        current.getLeft().setParent(current.getParent());
+        this._count -= 1;
+        return true;
+      }
+
+      // case 3: node has a right child and no left child
+      if (!current.hasLeft()) {
+        if (current.isRoot()) {
+          this._root = current.getRight();
+        } else if (k < current.getParent().getKey()) {
+          current.getParent().setLeft(current.getRight()).updateHeight();
+        } else {
+          current.getParent().setRight(current.getRight()).updateHeight();
+        }
+        current.getRight().setParent(current.getParent());
+        this._count -= 1;
+        return true;
+      }
+
+      // case 4: node has left and right children
+      const minRight = this.min(current.getRight());
+      current.setKey(minRight.getKey()).setValue(minRight.getValue());
+      return removeRecursively(minRight.getKey(), minRight);
+    };
+
+    return removeRecursively(key, this._root);
   }
 }