+++ src/avlTreeNode.js
@@ -1,5 +1,5 @@
 /**
- * @datastructures-js/binary-search-tree
+ * datastructures-js/binary-search-tree
  * @copyright 2020 Eyas Ranjous <eyas.ranjous@gmail.com>
  * @license MIT
  */
@@ -17,15 +17,16 @@ class AvlTreeNode extends BinarySearchTreeNode {
   }
 
   /**
-   * @internal
-   * rotates left (counter-clockwise) and updates parent and children
+   * Rotate-self left (counter-clockwise)
+   * @public
+   * @returns {AvlTreeNode}
    */
   rotateLeft() {
-    const right = this.getRight(); // this._right will be re-assigned
+    const right = this._right; // this._right will be re-assigned
 
     // set the node as a left child of its right child
     if (right !== null) {
-      if (right.getLeft() !== null) {
+      if (right.hasLeft()) {
         right.getLeft().setParent(this);
       }
 
@@ -37,7 +38,7 @@ class AvlTreeNode extends BinarySearchTreeNode {
     }
 
     // rebase parent's child to node's right child
-    if (this._parent !== null && right !== null) {
+    if (this.hasParent() && right !== null) {
       if (this._parent.getKey() < right.getKey()) {
         this._parent.setRight(right);
       } else {
@@ -49,33 +50,36 @@ class AvlTreeNode extends BinarySearchTreeNode {
     this._parent = right;
 
     this.updateHeight();
-    if (this._parent !== null) {
+    if (this.hasParent()) {
       this._parent.updateHeight();
     }
+
+    return this;
   }
 
   /**
-   * @internal
-   * rotates right (clockwise) and updates parent and children
+   * Rotate-self right (clockwise)
+   * @public
+   * @returns {AvlTreeNode}
    */
   rotateRight() {
-    const left = this.getLeft(); // this._left will be re-assigned
+    const left = this._left; // this._left will be re-assigned
 
     // set the node as a right child of its left child
     if (left !== null) {
-      if (left.getRight() !== null) {
+      if (left.hasRight()) {
         left.getRight().setParent(this);
       }
 
-      // rebase right child to node's right left child.
+      // rebase left child to node's left right child.
       this._left = left.getRight();
 
       left.setRight(this);
       left.setParent(this._parent);
     }
 
-    // rebase parent to node's left child
-    if (this._parent !== null && left !== null) {
+    // rebase parent's child to node's left child
+    if (this.hasParent() && left !== null) {
       if (this._parent.getKey() > left.getKey()) {
         this._parent.setLeft(left);
       } else {
@@ -83,35 +87,41 @@ class AvlTreeNode extends BinarySearchTreeNode {
       }
     }
 
-    // rebase parent to node's right child
+    // rebase parent to node's left child
     this._parent = left;
 
     this.updateHeight();
-    if (this._parent !== null) {
+    if (this.hasParent()) {
       this._parent.updateHeight();
     }
+
+    return this;
   }
 
   /**
-   * @internal
-   * rotates left child to left then itself to right
+   * Rotate-self to right after rotating left child to left
+   * @public
+   * @returns {AvlTreeNode}
    */
   rotateLeftRight() {
-    if (this._left !== null) {
+    if (this.hasLeft()) {
       this._left.rotateLeft();
     }
     this.rotateRight();
+    return this;
   }
 
   /**
-   * @internal
-   * rotates right child to right then itself to left
+   * Rotate-self to left after rotating right child to right
+   * @public
+   * @returns {AvlTreeNode}
    */
   rotateRightLeft() {
-    if (this._right !== null) {
+    if (this.hasRight()) {
       this._right.rotateRight();
     }
     this.rotateLeft();
+    return this;
   }
 
   /**
@@ -119,7 +129,7 @@ class AvlTreeNode extends BinarySearchTreeNode {
    * @return {number}
    */
   getLeftHeight() {
-    return this._left !== null ? this._left.getHeight() : 0;
+    return this.hasLeft() ? this.getLeft().getHeight() : 0;
   }
 
   /**
@@ -127,15 +137,17 @@ class AvlTreeNode extends BinarySearchTreeNode {
    * @return {number}
    */
   getRightHeight() {
-    return this._right !== null ? this._right.getHeight() : 0;
+    return this.hasRight() ? this.getRight().getHeight() : 0;
   }
 
   /**
-   * @internal
-   * updates the height of a node as the max height of its children
-   */
+   * Updates self height based on the max height of children
+   * @public
+   * @returns {AvlTreeNode}
+  */
   updateHeight() {
     this._height = Math.max(this.getLeftHeight(), this.getRightHeight()) + 1;
+    return this;
   }
 
   /**
@@ -147,12 +159,22 @@ class AvlTreeNode extends BinarySearchTreeNode {
   }
 
   /**
+   * Gets the balance of a node as the diff between left & right heights
    * @public
-   * calculate the balance of a node as the diff between left & right heights
    * @return {number}
    */
-  calculateBalance() {
+  getBalance() {
     return this.getLeftHeight() - this.getRightHeight();
+  }
+
+  /**
+   * Checks if the node is balanced
+   * @public
+   * @return {boolean}
+   */
+  isBalanced() {
+    const balance = this.getBalance();
+    return balance >= -1 && balance <= 1;
   }
 }