+++ src/binarySearchTreeNode.js
@@ -1,5 +1,5 @@
 /**
- * @datastructures-js/binary-search-tree
+ * datastructures-js/binary-search-tree
  * @copyright 2020 Eyas Ranjous <eyas.ranjous@gmail.com>
  * @license MIT
  */
@@ -17,11 +17,13 @@ class BinarySearchTreeNode {
   }
 
   /**
-   * @internal
+   * @public
    * @param {number|string}
+   * @returns {BinarySearchTreeNode}
    */
   setKey(key) {
     this._key = key;
+    return this;
   }
 
   /**
@@ -34,26 +36,34 @@ class BinarySearchTreeNode {
 
   /**
    * @public
-   * @param {object}
+   * @param {any} value
+   * @returns {BinarySearchTreeNode}
    */
   setValue(value) {
     this._value = value;
+    return this;
   }
 
   /**
    * @public
-   * @return {object}
+   * @return {any}
    */
   getValue() {
     return this._value;
   }
 
   /**
-   * @internal
-   * @param {BinarySearchTreeNode}
+   * @public
+   * @param {BinarySearchTreeNode|null} left
+   * @returns {BinarySearchTreeNode}
    */
   setLeft(left) {
-    this._left = left;
+    if (left && !(left instanceof BinarySearchTreeNode)) {
+      throw new Error('setLeft expects a BinarySearchTreeNode or null');
+    }
+
+    this._left = left || null;
+    return this;
   }
 
   /**
@@ -65,11 +75,25 @@ class BinarySearchTreeNode {
   }
 
   /**
-   * @internal
-   * @param {BinarySearchTreeNode}
+   * @public
+   * @return {boolean}
+   */
+  hasLeft() {
+    return this._left instanceof BinarySearchTreeNode;
+  }
+
+  /**
+   * @public
+   * @param {BinarySearchTreeNode|null} right
+   * @returns {BinarySearchTreeNode}
    */
   setRight(right) {
-    this._right = right;
+    if (right && !(right instanceof BinarySearchTreeNode)) {
+      throw new Error('setRight expects a BinarySearchTreeNode or null');
+    }
+
+    this._right = right || null;
+    return this;
   }
 
   /**
@@ -81,11 +105,25 @@ class BinarySearchTreeNode {
   }
 
   /**
-   * @internal
-   * @param {BinarySearchTreeNode}
+   * @public
+   * @return {boolean}
+   */
+  hasRight() {
+    return this._right instanceof BinarySearchTreeNode;
+  }
+
+  /**
+   * @public
+   * @param {BinarySearchTreeNode} parent
+   * @returns {BinarySearchTreeNode}
    */
   setParent(parent) {
-    this._parent = parent;
+    if (parent && !(parent instanceof BinarySearchTreeNode)) {
+      throw new Error('setParent expects a BinarySearchTreeNode or null');
+    }
+
+    this._parent = parent || null;
+    return this;
   }
 
   /**
@@ -94,6 +132,30 @@ class BinarySearchTreeNode {
    */
   getParent() {
     return this._parent;
+  }
+
+  /**
+   * @public
+   * @return {boolean}
+   */
+  hasParent() {
+    return this._parent instanceof BinarySearchTreeNode;
+  }
+
+  /**
+   * @public
+   * @return {boolean}
+   */
+  isRoot() {
+    return this._parent === null;
+  }
+
+  /**
+   * @public
+   * @return {boolean}
+   */
+  isLeaf() {
+    return !this.hasLeft() && !this.hasRight();
   }
 }